Quest Stuff to make tutorials and directions
I like an idiot copied over my stuff i was working on and had to go
through remaking it. going to remake my level over the weekend

the questLog is minorly functional including a questObject to add quest,
questSteps. needs functionality for incrementing quest and moving quest
to certain states.

also made the C# timer. it the same as the javascript just with
different function wrappers.

I need to still make the Respawn controller and objects

all the quest items are now in a subfolder of scripts to help arrange
them